requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
e2b>=0.14.0
openai>=1.0.0
httpx>=0.24.0
playwright>=1.35.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
